ai_serving/encoders/BiEncoders/requirements.txt
transformers
pytorch
fastapi
uvicorn